refactor: authentication views
