+++ smarter/smarter/apps/account/views/authentication/__init__.py
@@ -1,0 +1,25 @@
+"""
+Django Account Authentication views.
+"""
+
+from .account_views import (
+    AccountActivateView,
+    AccountActivationEmailView,
+    AccountDeactivateView,
+    AccountInactiveView,
+    AccountRegisterView,
+)
+from .login_view import LoginView
+from .logout_view import LogoutView
+from .social_auth import SocialAuthAlreadyAssociatedView
+
+__all__ = [
+    "AccountInactiveView",
+    "AccountRegisterView",
+    "AccountActivateView",
+    "AccountActivationEmailView",
+    "AccountDeactivateView",
+    "LoginView",
+    "LogoutView",
+    "SocialAuthAlreadyAssociatedView",
+]
